Update com.github.macadmins.Nudge.plist (#761)
* Update com.github.macadmins.Nudge.plist

- Updated `pfm_macos_min` because the latest Nudge app has a macOS 12 requirement [source](https://github.com/macadmins/nudge/wiki/v2.0-features#breaking-changes)
- Changed `pfm_default` for `Random Delay` to `true` to match Nudge 2.0.2+ behavior [source](https://github.com/macadmins/nudge/releases/tag/v2.0.2.81720)
- Removed `pfm_value_placeholder` for `Random Delay` as it should no longer be needed.

* Update com.github.macadmins.Nudge.plist

Reverted pfm_macos_min to v11

Reviewed Ari's comment rregarding `nudgeMajorUpgradeEventLaunch`. I did a github search on the Nudge repository, and determined that this is not a key, it's a mistake in documentation. `nudgeMajorUpgradeEventLaunchDelay` (notice "Delay") and `nudgeMinorUpdateEventLaunchDelay` are new keys.

These keys already exist in the Nudge manifext, however they were NOT included in `pfm_segments > UX` like other keys of this type. So I added them there.